(1682) harmonisation entre 1682 et 1686
