+++ packages/api/server/models/questionModel/serializeQuestion.ts
@@ -9,6 +9,7 @@ export default question => Object.assign(
         createdAt: question.questionCreatedAt !== null ? (question.questionCreatedAt.getTime() / 1000) : null,
         updatedAt: question.questionUpdatedAt !== null ? (question.questionUpdatedAt.getTime() / 1000) : null,
         solvedAt: question.questionSolvedAt !== null ? (question.questionSolvedAt.getTime() / 1000) : null,
+        answers: [],
         createdBy: {
             id: question.questionCreatedBy,
             first_name: question.userFirstName,
